rename discovery -> instance registry package
