+++ jfix-zookeeper/src/main/kotlin/ru/fix/zookeeper/instance/registry/InstanceIdGenerator.kt
@@ -1,4 +1,4 @@
-package ru.fix.zookeeper.discovery
+package ru.fix.zookeeper.instance.registry
 
 interface InstanceIdGenerator {
     fun nextId(instanceIds: List<String>) : String
