+++ jfix-zookeeper/src/main/kotlin/ru/fix/zookeeper/instance/registry/InstanceIdValidator.kt
@@ -1,4 +1,4 @@
-package ru.fix.zookeeper.discovery
+package ru.fix.zookeeper.instance.registry
 
 class InstanceIdValidator(private val maxInstancesCount: Int) {
